#126 자신이 작성한 리뷰 및 포스트에는 좋아요 누르지 못하도록 처리

+++ api/src/main/java/com/hot6/phopa/api/domain/review/service/ReviewApiService.java
@@ -142,8 +142,11 @@ public class ReviewApiService {
 
     public void like(Long reviewImageId) {
         UserDTO userDTO = PrincipleDetail.get();
-        UserEntity userEntity = userDTO.getId() != null ? userService.findById(userDTO.getId()) : null;
+        UserEntity userEntity = userService.findById(userDTO.getId());
         ReviewImageEntity reviewImageEntity = reviewService.getReviewImageById(reviewImageId);
+        if(reviewImageEntity.getReview().getUser().getId().equals(userEntity.getId())){
+            throw new SilentApplicationErrorException(ApplicationErrorType.CANNOT_BE_CREATED_USER);
+        }
         ReviewImageLikeEntity reviewImageLikeEntity = reviewService.getReviewImageLikeByReviewImageIdAndUserId(reviewImageEntity.getId(), userEntity.getId());
         if (reviewImageLikeEntity != null) {
             reviewService.deleteReviewImageLike(reviewImageLikeEntity);
